+++ crates/rslint_cli/CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-### Added 
+### Added
 
 - Added ways of configuring the formatter used through CLI and config
 - Made all fields of config public
+- Added the `rules` subcommand to show all available rules
 
 ## [0.2.1] - 2020-10-21